clean up workers better
note: there's the issue if you have outstanding requests what happens?
Terminate should terminate the worker but ... what about places in the code
stuck on `await`. That's up to you. Wait for your stuff to end for now.

Maybe should block until done but a really large file might block too
long and that might be the point of calling terminate.

Could add a flag, force = true, or something